chat_service/requirements.txt
Django>=5.1.7,<5.2
djangorestframework>=3.15.2,<3.16
channels==4.0.0
channels-redis==4.0.0
psycopg2-binary>=2.9.10,<2.10
pika>=1.3.2,<1.4
python-decouple==3.8
requests
Pillow>=11.1.0,<11.2
django-cors-headers>=4.7.0,<4.8
PyJWT==2.8.0
websocket-client
msgpack==1.0.8